+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Changed
 - Test step parameters handling
+- Mime type processing for data embedding was improved
 ### Fixed
 - Manually-reported nested steps now correctly fail all parents